Add Sharing button
Add Sharing button which entails using an Intent which can be used
explicit or generic. You used an example of the explicit type when you
specificqally defined the Intent that launches your app, which the
manifest then identifies as MainActivity.

In menu/main.xml we add <item> with
ActionProviderClass="android.widget.ShareActionProvider" and then we use
it in our Java code.

Contact: joaquin@isla.io